# Move Mapgen V7 river generation into the main generation loop