All terrain generation now occurs in one loop, instead of rivers
being carved afterwards in a separate loop. This fixes the removal
of nodes added by mods in 'register on generated'.

Also fixes a bug where rivergen is disabled at floatland altitudes
even when floatlands are disabled.
/////////////////////////////////////////////////

Closes #7878 and closes #10638 
Having all terrain generation in one loop makes a mapgen simpler, more robust and reduces the chances of bugs. I also expect a slight performance improvement due to not needing a second generation loop.

Identical terrain test
-
I mapgen-aliased water to air in MTG and compared a map of master branch to a map of this PR.
Area was chosen to have a river channel cutting through mountains for the most thorough testing.
Config:
```
fixed_map_seed = 0
static_spawnpoint = 525, 4.5, -445
mg_flags = nodecorations,nocaves,nobiomes,nodungeons,light,noores
```
In game: `/emergeblocks (0, -32, -1000) (1000, 127, 0)`
MT mapper: `--min-y -32 --max-y 127 --geometry 100:-900+800+800`
Resulting maps are below, save and open both, switch rapidly between to check they are identical.

![rivtestair_master_100_-900_800_800](https://user-images.githubusercontent.com/3686677/99569128-28d24900-29c8-11eb-977d-25d3146bda4f.png)

^ Master branch

![rivtestair_PR](https://user-images.githubusercontent.com/3686677/99569175-35ef3800-29c8-11eb-97c6-7ecaf5a5eb78.png)

^ PR

Test bug is fixed
-
I confirmed #10638 is fixed.
Config that creates extreme height and solid mountain terrain:
```
fixed_map_seed = 0
static_spawnpoint = 525, 4.5, -445
mgv7_np_mount_height = {
	flags = eased
	lacunarity = 2
	offset = 10000
	scale = 0
	spread = (1000,1000,1000)
	seed = 72449
	octaves = 3
	persistence = 0.6
}
mgv7_np_mountain = {
	flags = 
	lacunarity = 2
	offset = 1
	scale = 0.1
	spread = (250,350,250)
	seed = 5333
	octaves = 5
	persistence = 0.63
}
```

Chunk generation time test
-
I used the commented-out timing lines in code to compare chunk generation times of PR and master. Spawning the player near a river as before.
Config:
```
max_block_generate_distance = 1
fixed_map_seed = 0
static_spawnpoint = 525, 4.5, -445
mg_flags = nodecorations,nocaves,nobiomes,nodungeons,light,noores
```
Results:
```
Master

makeChunk: 62ms
makeChunk: 51ms
makeChunk: 70ms
makeChunk: 64ms

PR

makeChunk: 60ms
makeChunk: 50ms
makeChunk: 67ms
makeChunk: 63ms
```
Surprisingly noticeable improvement of 1-3ms less.

No river channel in floatland test
-
Floatland density was increased to be almost solid. Floatlands were confirmed to not have river canyons cut into them.
Config:
```
mgv7_spflags = mountains,ridges,floatlands,caverns
fixed_map_seed = 0
static_spawnpoint = 525, 4.5, -445
mgv7_floatland_density = 1
```