@paramat 2020-11-19
> Separate processing steps would be easier to understand than everything merged into generateTerrain. Not much we could do about that, right?

Correct.
The code is only slightly less easy to understand, but is a better implementation, fixes bugs and avoids future problems, and (see new test in first post) significantly improves performance. So certainly an improvement.
River generation is now done with per-node checks in the main loop, just like mountain generation, so this is familiar and consistent.

I did some more tests, see first post, and am satisfied with the tests now. No longer WIP.